type refactor to another separate file
