+++ src/context/DataSource.context.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, createContext } from 'react';
 import axios from 'axios';
-import { Flight } from '../types';
+import { Flight } from '../@types/flight';
 import { FLIGHTS_ALL_API } from '../utils/endpoints';
 
 type TDataSourceProvider = {
